backend/requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
sqlalchemy==2.0.45
pymysql==1.1.2
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.21
pydantic==2.12.5
email-validator
opencv-python==4.12.0.88
numpy==2.2.6
tensorflow==2.20.0
pillow
cryptography     